move m2m tests to their own module
